move the Dashboard pages to the :ui:base module so they can be linked to from other UIs
