--- app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/AllFavoritesScreen.kt
@@ -1,7 +1,0 @@
-package org.cru.godtools.ui.dashboard.home
-
-import com.slack.circuit.runtime.screen.Screen
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-object AllFavoritesScreen : Screen
