--- app/src/main/kotlin/org/cru/godtools/ui/dashboard/lessons/LessonsScreen.kt
@@ -1,7 +1,0 @@
-package org.cru.godtools.ui.dashboard.lessons
-
-import com.slack.circuit.runtime.screen.Screen
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data object LessonsScreen : Screen
