+++ ui/base/src/main/kotlin/org/cru/godtools/base/ui/circuit/screen/dashboard/page/ToolsScreen.kt
@@ -1,0 +1,7 @@
+package org.cru.godtools.base.ui.circuit.screen.dashboard.page
+
+import kotlinx.parcelize.Parcelize
+import org.cru.godtools.base.ui.dashboard.Page
+
+@Parcelize
+data object ToolsScreen : DashboardPage(Page.ALL_TOOLS)
